backend/requirements.txt
# VisionPhase ML Backend Dependencies
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18

# ML - Segmentation
segment-anything>=1.0
torch>=2.0.0
torchvision>=0.15.0

# ML - Classification & Generation
transformers>=4.40.0
diffusers>=0.30.0
accelerate>=0.30.0
safetensors>=0.4.0
peft>=0.10.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.12.0
scikit-image>=0.22.0

# Optional - Memory Optimization
xformers>=0.0.25

# Optional - Super Resolution
# realesrgan
# basicsr
